Moved Model testing into their own folder instead in repoTesting
